Add files via upload
Developed in Android Studio with Java, this app features a comprehensive list of all American presidents, complete with profiles that include their photos and years in office. Users can add new presidents to the list, creating custom profiles that are stored in a local database. Each entry includes details like an image and term years, making it easy to browse and learn more about the nation's leaders throughout history.